+++ tailwind.config.js
@@ -1,4 +1,5 @@
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
   presets: [
     require('./src/tailwindcss.js')
   ],